move contract methods api call inside own component
